# Cache fromRegex patterns