--- a/benchmarks/bench.hs
+++ b/benchmarks/bench.hs
@@ -4,10 +4,14 @@ import           Criterion.Main
 import           Data.Aeson     (Value (..))
 import           Expr           (Expr (..))
 import           Fake           (eval, runFakeT)
+import Control.Monad (replicateM)
 
 
 exec :: Expr -> IO Value
 exec expr = runFakeT (Just 1) (eval expr)
+
+execRepeated :: Int -> Expr -> IO [Value]
+execRepeated num expr = runFakeT (Just 1) (replicateM num (eval expr))
 
 
 main :: IO ()
@@ -16,4 +20,5 @@ main = defaultMain
                  , bench "uuid4" $ nfIO (exec "uuid4") ]
   , bgroup "numbers" [ bench "randomInt-0-100"  $ nfIO (exec "randomInt(0, 100)" )
                      , bench "randomInt-0-1000" $ nfIO (exec "randomInt(0, 1000)" ) ]
+  , bgroup "regex" [ bench "from-regex" $ nfIO (execRepeated 20 "fromRegex('[0-1]{10}')") ]
   ]