Without caching:

    benchmarking regex/from-regex
    time                 281.5 μs   (281.2 μs .. 281.8 μs)
                         1.000 R²   (1.000 R² .. 1.000 R²)
    mean                 280.1 μs   (279.6 μs .. 281.2 μs)
    std dev              2.518 μs   (1.417 μs .. 4.378 μs)

With:

    benchmarking regex/from-regex
    time                 141.7 μs   (141.5 μs .. 141.9 μs)
                         1.000 R²   (1.000 R² .. 1.000 R²)
    mean                 141.8 μs   (141.6 μs .. 142.0 μs)
    std dev              565.6 ns   (475.8 ns .. 676.3 ns)